\name{cowsgopoo}
\alias{cowsgopoo}
\title{Function for extracting time in each age class for calculating nitrogen 
deposition in grazed fields}
\usage{
cowsgopoo(data, ID, DOB, )
}
\arguments{
  \item{data}{data frame of class 'data.frame'}

  \item{ID}{column name of the variable containing the identifier/ID for each 
  individual animal}

  \item{ID}{column name of the variable containing the identifier/ID for each 
  individual animal}

  \item{ID}{column name of the variable containing the identifier/ID for each 
  individual animal}

  \item{ID}{column name of the variable containing the identifier/ID for each 
  individual animal}
}
\value{
Data frame
}
\description{
cowsgopoo
}
\examples{

}

